requirements.txt
langchain
langchain-chroma
langchain-huggingface
chromadb
sentence-transformers
pypdf
httpx
python-dotenv
ipykernel
pytest
